fix overflow

+++ util.go
@@ -4,7 +4,7 @@ import "math"
 
 // Min returns the minimum number of passed int slices.
 func Min(is ...int) int {
-	min := int(math.MaxInt64)
+	min := int64(math.MaxInt64)
 	for _, v := range is {
 		if min > v {
 			min = v
